+++ Backend/CoffeeBook/Controllers/DiscountController.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -51,35 +52,71 @@ namespace CoffeeBook.Controllers
         [HttpPost]
         public ActionResult Post(Discount discount)
         {
-            var result = service.save(discount);
-            if (result > 0)
-                return Ok();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1" || Role == "2")
+                {
+                    var result = service.save(discount);
+                    if (result > 0)
+                        return Ok();
 
-            return BadRequest();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
         }
 
-        [Route("discount/edit")]
+        [Route("discount/edit/{id}")]
         [HttpPut]
         public ActionResult Put(int id, Discount discount)
         {
-            var result = service.Update(id, discount);
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1" || Role == "2")
+                {
+                    var result = service.Update(id, discount);
 
-            if (result > 0)
-                return Ok();
+                    if (result > 0)
+                        return Ok();
 
-            return BadRequest();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
         }
 
         [Route("discount/delete/{id}")]
         [HttpDelete]
         public ActionResult Delete(int id)
         {
-            var result = service.DeleteById(id);
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1" || Role == "2")
+                {
+                    var result = service.DeleteById(id);
 
-            if (result > 0)
-                return Ok();
+                    if (result > 0)
+                        return Ok();
 
-            return BadRequest();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
+        }
+
+        private string getCurrentRole(string jwt)
+        {
+            var handler = new JwtSecurityTokenHandler();
+            var jsonToken = handler.ReadToken(jwt);
+            var tokenS = jsonToken as JwtSecurityToken;
+            var role = tokenS.Claims.First(claim => claim.Type == "RoleId").Value;
+            return role;
         }
     }
 }
