+++ Backend/CoffeeBook/Controllers/SupplierController.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -49,27 +50,63 @@ namespace CoffeeBook.Controllers
         [HttpPost]
         public ActionResult Post(Supplier supplier)
         {
-            int res = service.save(supplier);
-            if (res > 0) return Ok();
-            return BadRequest();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1")
+                {
+                    int res = service.save(supplier);
+                    if (res > 0) return Ok();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
         }
 
         [Route("supplier/edit/{id}")]
         [HttpPut]
         public ActionResult Put(int id,Supplier supplier)
         {
-            int res = service.Update(id, supplier);
-            if (res > 0) return Ok();
-            else return BadRequest();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1")
+                {
+                    int res = service.Update(id, supplier);
+                    if (res > 0) return Ok();
+                    else return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
         }
 
         [Route("supplier/delete/{id}")]
         [HttpDelete]
         public ActionResult Delete(int id)
         {
-            int res = service.deleteById(id);
-            if (res > 0) return Ok();
-            return BadRequest();
+            string jwt = Request.Cookies["jwt"];
+            if (!string.IsNullOrEmpty(jwt))
+            {
+                var Role = getCurrentRole(jwt);
+                if (Role == "1")
+                {
+                    int res = service.deleteById(id);
+                    if (res > 0) return Ok();
+                    return BadRequest();
+                }
+            }
+            return Unauthorized(new { message = "Bạn không có quyền truy cập" });
+        }
+
+        private string getCurrentRole(string jwt)
+        {
+            var handler = new JwtSecurityTokenHandler();
+            var jsonToken = handler.ReadToken(jwt);
+            var tokenS = jsonToken as JwtSecurityToken;
+            var role = tokenS.Claims.First(claim => claim.Type == "RoleId").Value;
+            return role;
         }
     }
 }
